+++ PluginInfo.cs
@@ -4,6 +4,6 @@ namespace BetterTames
     {
         public const string PLUGIN_NAME = "BetterTames";
         public const string PLUGIN_GUID = "xtav1m." + PLUGIN_NAME;
-        public const string PLUGIN_VERSION = "0.8.0";
+        public const string PLUGIN_VERSION = "0.9.0";
     }
 }